# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quest API
  description: |
    REST API for Quest - a Jira-like project and task management system.

    ## Authentication
    All endpoints require authentication via session cookie (NextAuth.js) or Bearer token.

    ## Base URL
    - Production: `https://quest-manas-pradhans-projects.vercel.app`
    - Development: `http://localhost:3000`

    ## Notes for Mobile Developers
    This API specification documents the full capabilities of Quest. All endpoints are
    implemented REST API routes that wrap Next.js Server Actions, making them consumable
    by iOS, Android, and other platforms via standard HTTP requests.
  version: 1.0.0
  contact:
    name: Quest Development Team
servers:
  - url: https://quest-manas-pradhans-projects.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Validation error or user exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Request password reset email
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset email sent (always returns success to prevent enumeration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/auth/password-reset:
    post:
      tags: [Auth]
      summary: Reset password with token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===========================================
  # User Profile
  # ===========================================
...
  /api/issues/{issueId}/subtasks:
    parameters:
      - $ref: '#/components/parameters/IssueId'

    get:
      tags: [Issues]
      summary: List subtasks for an issue
      operationId: listSubtasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of subtasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subtask'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Issues]
      summary: Create a subtask
      operationId: createSubtask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubtaskRequest'
      responses:
        '201':
          description: Subtask created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '400':
          description: Validation error or cannot create subtask of subtask
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/issues/{issueId}/comments:
    parameters:
      - $ref: '#/components/parameters/IssueId'

    get:
      tags: [Comments]
      summary: List comments on an issue
      operationId: listComments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Comments]
      summary: Add a comment to an issue
      operationId: createComment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/comments/{commentId}:
    parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: string

    patch:
      tags: [Comments]
      summary: Update a comment
      operationId: updateComment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
      responses:
        '200':
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Can only edit own comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Comments]
      summary: Delete a comment
      operationId: deleteComment
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only author, owner, or admin can delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================
  # Attachments (IMPLEMENTED)
  # ===========================================
...
  /api/filters/{filterId}:
    parameters:
      - name: filterId
        in: path
        required: true
        schema:
          type: string

    patch:
      tags: [Filters]
      summary: Update a saved filter
      operationId: updateFilter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFilterRequest'
      responses:
        '200':
          description: Filter updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilterResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Can only update own filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Filters]
      summary: Delete a saved filter
      operationId: deleteFilter
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Filter deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Can only delete own filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================
  # Bulk Operations
  # ===========================================
  /api/issues/bulk/status:
    post:
      tags: [Bulk Operations]
      summary: Bulk update issue status
      operationId: bulkUpdateStatus
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issueIds, status]
              properties:
                issueIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 50
                status:
                  $ref: '#/components/schemas/IssueStatus'
      responses:
        '200':
          description: Issues updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResponse'
        '400':
          description: Some issues not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: No permission for some issues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/issues/bulk/assign:
    post:
      tags: [Bulk Operations]
      summary: Bulk assign issues
      operationId: bulkAssign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issueIds]
              properties:
                issueIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 50
                assigneeId:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Issues assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/issues/bulk/priority:
    post:
      tags: [Bulk Operations]
      summary: Bulk update issue priority
      operationId: bulkUpdatePriority
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issueIds, priority]
              properties:
                issueIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 50
                priority:
                  $ref: '#/components/schemas/IssuePriority'
      responses:
        '200':
          description: Issues updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/issues/bulk/delete:
    post:
      tags: [Bulk Operations]
      summary: Bulk delete issues
      operationId: bulkDelete
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issueIds]
              properties:
                issueIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 50
      responses:
        '200':
          description: Issues deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT or session token
  parameters:
...
    IssueId:
      name: issueId
      in: path
      required: true
      schema:
        type: string
      description: Issue ID (CUID)
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not found"
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    BulkResponse:
      type: object
      properties:
        success:
          type: boolean
        count:
          type: integer

    # ===========================================
    # Auth Schemas
    # ===========================================
    SignUpRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Must contain lowercase, uppercase, and number
...
    ResetPasswordRequest:
      type: object
      required: [token, password, confirmPassword]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
        confirmPassword:
          type: string
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
        image:
          type: string
          nullable: true

    # ===========================================
    # Workspace Schemas
    # ===========================================
...
    IssueType:
      type: string
      enum: [EPIC, STORY, TASK, BUG]
    IssueStatus:
      type: string
      enum: [BACKLOG, TODO, IN_PROGRESS, IN_REVIEW, DONE, CANCELLED]
    IssuePriority:
      type: string
      enum: [URGENT, HIGH, MEDIUM, LOW, NONE]
    Issue:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
          description: Human-readable key (e.g., "QUEST-123")
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/IssueType'
        status:
          $ref: '#/components/schemas/IssueStatus'
        priority:
          $ref: '#/components/schemas/IssuePriority'
        projectId:
          type: string
        assigneeId:
          type: string
          nullable: true
        reporterId:
          type: string
        dueDate:
          type: string
          format: date-time
          nullable: true
        order:
          type: integer
        parentId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IssueWithRelations:
      allOf:
        - $ref: '#/components/schemas/Issue'
        - type: object
          properties:
            assignee:
              $ref: '#/components/schemas/UserSummary'
              nullable: true
            reporter:
              $ref: '#/components/schemas/UserSummary'
            labels:
              type: array
              items:
                type: object
                properties:
                  label:
                    $ref: '#/components/schemas/Label'
            subtasks:
              type: array
              items:
                $ref: '#/components/schemas/Subtask'
    Subtask:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        title:
          type: string
        type:
          $ref: '#/components/schemas/IssueType'
        status:
          $ref: '#/components/schemas/IssueStatus'
        priority:
          $ref: '#/components/schemas/IssuePriority'
    Label:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          description: Hex color code
...
    IssueResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/IssueWithRelations'
    CreateSubtaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        type:
          type: string
          enum: [TASK, BUG]
          default: TASK
        priority:
          $ref: '#/components/schemas/IssuePriority'
        assigneeId:
          type: string
          nullable: true

    # ===========================================
    # Comment Schemas
    # ===========================================
    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        issueId:
          type: string
        authorId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        author:
          $ref: '#/components/schemas/UserSummary'
    CommentResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Comment'

    # ===========================================
    # Attachment Schemas
    # ===========================================
...
    IssueFilters:
      type: object
      properties:
        search:
          type: string
        status:
          type: array
          items:
            $ref: '#/components/schemas/IssueStatus'
        type:
          type: array
          items:
            $ref: '#/components/schemas/IssueType'
        priority:
          type: array
          items:
            $ref: '#/components/schemas/IssuePriority'
        assigneeId:
          type: array
          items:
            type: string
        labelId:
          type: array
          items:
            type: string
    SavedFilter:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        filters:
          $ref: '#/components/schemas/IssueFilters'
        projectId:
          type: string
        userId:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateFilterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        filters:
          $ref: '#/components/schemas/IssueFilters'
        isDefault:
          type: boolean
    FilterResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SavedFilter'